refactor(server): Moved all auth logic from v1.py to new module auth.py
