+++ server/src/testflinger/api/v1.py
@@ -18,11 +18,9 @@
 import importlib.metadata
 import os
 import uuid
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timezone
 from http import HTTPStatus
 
-import bcrypt
-import jwt
 import requests
 from apiflask import APIBlueprint, abort
 from flask import jsonify, request, send_file
@@ -32,7 +30,7 @@ from urllib3.util.retry import Retry
 from werkzeug.exceptions import BadRequest
 
 from testflinger import database
-from testflinger.api import schemas
+from testflinger.api import auth, schemas
 
 jobs_metric = Counter(
     "jobs", "Number of jobs", ["queue"], namespace="testflinger"
@@ -101,123 +99,6 @@ def has_attachments(data: dict) -> bool:
     )
 
 
-def decode_jwt_token(auth_token: str, secret_key: str) -> dict:
-    """
-    Decode authorization token using the secret key. Aborts with
-    an HTTP error if it does not exist or if it fails to decode.
-    """
-    if auth_token is None:
-        abort(401, "Unauthorized")
-    try:
-        decoded_jwt = jwt.decode(
-            auth_token,
-            secret_key,
-            algorithms="HS256",
-            options={"require": ["exp", "iat", "sub"]},
-        )
-    except jwt.exceptions.ExpiredSignatureError:
-        abort(401, "Token has expired")
-    except jwt.exceptions.InvalidTokenError:
-        abort(403, "Invalid Token")
-
-    return decoded_jwt
-
-
-def check_token_priority(
-    auth_token: str, secret_key: str, queue: str, priority: int
-) -> bool:
-    """
-    Check if the requested priority is less than the max priority
-    specified in the authorization token if it exists.
-    """
-    if priority == 0:
-        return
-    decoded_jwt = decode_jwt_token(auth_token, secret_key)
-    permissions = decoded_jwt.get("permissions", {})
-    max_priority_dict = permissions.get("max_priority", {})
-    star_priority = max_priority_dict.get("*", 0)
-    queue_priority = max_priority_dict.get(queue, 0)
-    max_priority = max(star_priority, queue_priority)
-    if priority > max_priority:
-        abort(
-            403,
-            (
-                f"Not enough permissions to push to {queue}",
-                f"with priority {priority}",
-            ),
-        )
-
-
-def check_token_queue(auth_token: str, secret_key: str, queue: str):
-    """
-    Check if the queue is in the restricted list. If it is, then it
-    checks the authorization token for restricted queues the user is
-    allowed to use.
-    """
-    if not database.check_queue_restricted(queue):
-        return
-    decoded_jwt = decode_jwt_token(auth_token, secret_key)
-    permissions = decoded_jwt.get("permissions", {})
-    allowed_queues = permissions.get("allowed_queues", [])
-    if queue not in allowed_queues:
-        abort(
-            403,
-            (
-                "Not enough permissions to push to the ",
-                f"restricted queue: {queue}",
-            ),
-        )
-
-
-def check_token_reservation_timeout(
-    auth_token: str, secret_key: str, reservation_timeout: int, queue: str
-):
-    """
-    Check if the requested reservation is either less than the max
-    or that their token gives them the permission to use a higher one.
-    """
-    # Max reservation time defaults to 6 hours
-    max_reservation_time = 6 * 60 * 60
-    if reservation_timeout <= max_reservation_time:
-        return
-    decoded_jwt = decode_jwt_token(auth_token, secret_key)
-    permissions = decoded_jwt.get("permissions", {})
-    max_reservation_time_dict = permissions.get("max_reservation_time", {})
-    queue_reservation_time = max_reservation_time_dict.get(queue, 0)
-    star_reservation_time = max_reservation_time_dict.get("*", 0)
-    max_reservation_time = max(queue_reservation_time, star_reservation_time)
-    if reservation_timeout > max_reservation_time:
-        abort(
-            403,
-            (
-                f"Not enough permissions to push to {queue}",
-                f"with reservation timeout {reservation_timeout}",
-            ),
-        )
-
-
-def check_token_permissions(
-    auth_token: str,
-    secret_key: str,
-    job_data: dict,
-):
-    """
-    Validate token received from client and checks if it can
-    push a job to the queue with the requested priority.
-    """
-    priority_level = job_data.get("job_priority", 0)
-    job_queue = job_data["job_queue"]
-    check_token_priority(auth_token, secret_key, job_queue, priority_level)
-    check_token_queue(auth_token, secret_key, job_queue)
-
-    reserve_data = job_data.get("reserve_data", {})
-    # default reservation timeout is 1 hour
-    reservation_timeout = reserve_data.get("timeout", 3600)
-    check_token_reservation_timeout(
-        auth_token, secret_key, reservation_timeout, job_queue
-    )
-
-
 def job_builder(data: dict, auth_token: str):
     """Build a job from a dictionary of data."""
     job = {
@@ -242,7 +123,7 @@ def job_builder(data: dict, auth_token: str):
         data["attachments_status"] = "waiting"
 
     priority_level = data.get("job_priority", 0)
-    check_token_permissions(
+    auth.check_token_permissions(
         auth_token,
         os.environ.get("JWT_SIGNING_KEY"),
         data,
@@ -913,47 +794,6 @@ def get_jobs_by_queue(queue_name):
     return jsonify(jobs_in_queue)
 
 
-def generate_token(allowed_resources, secret_key):
-    """
-    Generate JWT token with queue permission given a secret key
-    See retrieve_token for more information on the contents of
-    the token payload.
-    """
-    expiration_time = datetime.now(timezone.utc) + timedelta(seconds=30)
-    token_payload = {
-        "exp": expiration_time,
-        "iat": datetime.now(timezone.utc),  # Issued at time
-        "sub": "access_token",
-        "permissions": allowed_resources,
-    }
-    token = jwt.encode(token_payload, secret_key, algorithm="HS256")
-    return token
-
-
-def validate_client_key_pair(client_id: str, client_key: str):
-    """
-    Check client_id and key pair for validity and
-    returns their permissions.
-    """
-    if client_key is None:
-        return None
-    client_key_bytes = client_key.encode("utf-8")
-    client_permissions_entry = database.mongo.db.client_permissions.find_one(
-        {
-            "client_id": client_id,
-        }
-    )
-
-    if client_permissions_entry is None or not bcrypt.checkpw(
-        client_key_bytes,
-        client_permissions_entry["client_secret_hash"].encode("utf8"),
-    ):
-        return None
-    client_permissions_entry.pop("_id", None)
-    client_permissions_entry.pop("client_secret_hash", None)
-    return client_permissions_entry
-
-
 @v1.post("/oauth2/token")
 def retrieve_token():
     """
@@ -982,9 +822,9 @@ def retrieve_token():
             401,
         )
 
-    allowed_resources = validate_client_key_pair(client_id, client_key)
+    allowed_resources = auth.validate_client_key_pair(client_id, client_key)
     if allowed_resources is None:
         return "Invalid client id or client key", 401
     secret_key = os.environ.get("JWT_SIGNING_KEY")
-    token = generate_token(allowed_resources, secret_key)
+    token = auth.generate_token(allowed_resources, secret_key)
     return token
